New version bugfixes (#70)
* bugfixes

* Update data_quality_tables.sql

* Update pyproject.toml

* Unit test adjustments for code changes

---------

+++ scripts/data_quality_tables.sql
@@ -82,7 +82,7 @@ CREATE TABLE IF NOT EXISTS ${catalog}.data_quality.validatie (
   regelId STRING,
   aantalValideRecords BIGINT,
   aantalReferentieRecords BIGINT,
-  percentageValideRecords INT,
+  percentageValideRecords DOUBLE,
   dqDatum TIMESTAMP,
   dqResultaat STRING)
 USING delta
